requirements.txt
python-dotenv==1.0.0
requests==2.31.0  
pandas==2.1.4 
numpy==1.26.0 
alpaca==0.8.0
alpaca-py==0.4.0